+++ internal/applications/messenger/pkg/client/builder.go
@@ -18,14 +18,6 @@ func NewBuilder() IBuilder {
 	return &sBuilder{}
 }
 
-func (p *sBuilder) PingMessage() hls_request.IRequest {
-	return hls_request.NewRequestBuilder().
-		WithMethod(http.MethodGet).
-		WithHost(hlm_settings.CServiceFullName).
-		WithPath(hlm_settings.CPingPath).
-		Build()
-}
-
 func (p *sBuilder) PushMessage(pBody []byte) hls_request.IRequest {
 	return hls_request.NewRequestBuilder().
 		WithMethod(http.MethodPost).